add a small changes in work flow
