+++ src/main/java/ru/bprn/printhouse/views/template/SelectMaterailsDialog.java
@@ -48,9 +48,73 @@ public class SelectMaterailsDialog extends Dialog {
         this.add(layout);
     }
 
-
-
     public void setSelectedMaterial(Set<Material> materials) {
         if (!materials.isEmpty()) for (Material m : materials) grid.select(m);
     }
+/*
+    private void addMaterialSection() {
+
+        var typeOfMaterialCombo = new ComboBox<TypeOfMaterial>();
+        var thicknessCombo = new ComboBox<Thickness>();
+        var sizeOfPrintLeafCombo = new ComboBox<SizeOfPrintLeaf>();
+        var grid = new Grid<>(Material.class, false);
+
+        typeOfMaterialCombo.setItems(typeOfMaterialService.findAll());
+        typeOfMaterialCombo.setAllowCustomValue(false);
+        typeOfMaterialCombo.addValueChangeListener(e->{
+            if (e.getValue()!=null) {
+                thicknessCombo.setItems(materialService.findAllThicknessByTypeOfMaterial(e.getValue()));
+                comboBoxViewFirstElement(thicknessCombo);
+                sizeOfPrintLeafCombo.setItems(materialService.findAllSizeOfPrintLeafByTypeOfMaterial(e.getValue()));
+                comboBoxViewFirstElement(sizeOfPrintLeafCombo);
+            }
+            grid.setItems(materialService.findByFilters(e.getValue(), sizeOfPrintLeafCombo.getValue(), thicknessCombo.getValue()));
+        });
+
+        var listThickness = materialService.findAllThicknessByTypeOfMaterial(typeOfMaterialCombo.getValue());
+        if (listThickness!=null) thicknessCombo.setItems(listThickness);
+        thicknessCombo.setAllowCustomValue(false);
+        thicknessCombo.addValueChangeListener(e->{
+            grid.setItems(materialService.findByFilters(typeOfMaterialCombo.getValue(), sizeOfPrintLeafCombo.getValue(), e.getValue()));
+        });
+
+        var listSizeOfPrintLeaf = materialService.findAllSizeOfPrintLeafByTypeOfMaterial(typeOfMaterialCombo.getValue());
+        if (listSizeOfPrintLeaf!=null) sizeOfPrintLeafCombo.setItems(listSizeOfPrintLeaf);
+        sizeOfPrintLeafCombo.setAllowCustomValue(false);
+        sizeOfPrintLeafCombo.addValueChangeListener(e->{
+            grid.setItems(materialService.findByFilters(typeOfMaterialCombo.getValue(), e.getValue(), thicknessCombo.getValue()));
+        });
+
+        grid.addColumn(Material::getName).setHeader("Название");
+        Grid.Column<Material> typeColumn = grid.addColumn(Material::getTypeOfMaterial).setHeader("Тип материала");
+        Grid.Column<Material> sizeColumn = grid.addColumn(Material::getSizeOfPrintLeaf).setHeader("Размер печатного листа");
+        Grid.Column<Material> thicknessColumn = grid.addColumn(Material::getThickness).setHeader("Плотность");
+        grid.setSelectionMode(Grid.SelectionMode.SINGLE);
+        grid.addThemeVariants(GridVariant.LUMO_ROW_STRIPES);
+        grid.setHeight("270px");
+        grid.setItems(materialService.findByFilters(typeOfMaterialCombo.getValue(), sizeOfPrintLeafCombo.getValue(), thicknessCombo.getValue()));
+
+        //templateBinder.forField(grid.).bind(WorkFlow::getMaterial, WorkFlow::setMaterial);
+        templateBinder.forField(grid.asSingleSelect()).
+                withValidator(Objects::nonNull, "Выделите строку!")
+                .bind(WorkFlow::getMaterial, WorkFlow::setMaterial);
+
+        grid.getHeaderRows().clear();
+        HeaderRow headerRow = grid.appendHeaderRow();
+
+        headerRow.getCell(typeColumn).setComponent(typeOfMaterialCombo);
+        headerRow.getCell(sizeColumn).setComponent(sizeOfPrintLeafCombo);
+        headerRow.getCell(thicknessColumn).setComponent(thicknessCombo);
+
+        grid.addSelectionListener(selectionEvent -> {
+            if (autoNamed.getValue()) {
+                templateBinder.getBean().setName(setDefaultName());
+                templateBinder.refreshFields();
+            }
+        });
+
+        this.add(grid);
+    }
+
+ */
 }
